requirements.python_base.txt
# Networking & APIs
requests==2.31.0
httpx==0.27.0
paho-mqtt==1.6.1

# MongoDB
pymongo==4.6.3

# Utilities
setuptools==69.5.1
wheel==0.43.0
debugpy==1.8.1